+++ src/components/control/money.jsx
@@ -1,16 +1,17 @@
 import PropTypes from 'prop-types';
 import React, { forwardRef, useImperativeHandle, useState, useRef, useEffect } from 'react';
-import CustomFieldInputText from '../custom-field-input-text/custom-field-input-text.jsx';
-import CustomFieldInputNumber from '../custom-field-input-number/custom-field-input-number.jsx';
-import currencyCodeType from './currency-code-type.js';
-import currencyMetaData from './currency-meta-data.js';
+import Input from './input.jsx';
+import Number from './number.jsx';
+import currencyCodeType from '../custom-field-input-currency/currency-code-type.js';
+import currencyMetaData from '../custom-field-input-currency/currency-meta-data.js';
+import { initialInputValid, subunitToUnit, formatValue } from './money-formatter.js';
 import useForwardedRef from '../../hooks/use-forwarded-ref.js';
-import { initialInputValid, subunitToUnit, formatValue } from '../control/money-formatter.js';
 
-const CustomFieldInputCurrency = forwardRef(function CustomFieldInputCurrency(props, forwardedRef) {
+const Money = forwardRef(function Money(props, forwardedRef) {
   const [input, setInput] = useState(subunitToUnit(props.value, props.currencyCode));
-  const [isEditing, setIsEditing] = useState(false);
   const [isFocused, setIsFocused] = useState(false);
+  const [validationMessage, setValidationMessage] = useState(props.validationMessage);
+  const [isEditing, setIsEditing] = useState(!!validationMessage);
   const componentRef = useRef(null);
   const numberRef = useRef(null);
   const valueRef = isEditing ? numberRef : componentRef;
@@ -19,6 +20,7 @@ const CustomFieldInputCurrency = forwardRef(function CustomFieldInputCurrency(pr
   function handleOnBlur(event) {
     if (numberRef.current.validity.valid) {
       setInput(event.target.value === '' ? undefined : parseFloat(event.target.value));
+      setValidationMessage('');
       setIsEditing(false);
     }
 
@@ -32,8 +34,8 @@ const CustomFieldInputCurrency = forwardRef(function CustomFieldInputCurrency(pr
     setIsFocused(true);
   }
 
-  function onChange() {
-    props.onChange({ target: ref.current });
+  function onInvalid(event) {
+    setValidationMessage(event.detail.validationMessage);
   }
 
   useEffect(() => {
@@ -41,6 +43,10 @@ const CustomFieldInputCurrency = forwardRef(function CustomFieldInputCurrency(pr
       setIsEditing(true);
     }
   }, []);
+
+  useEffect(() => {
+    props.onInvalid({ detail: { validationMessage } });
+  }, [validationMessage]);
 
   useEffect(() => {
     if (isEditing && isFocused) {
@@ -54,8 +60,7 @@ const CustomFieldInputCurrency = forwardRef(function CustomFieldInputCurrency(pr
 
   useImperativeHandle(ref, () => ({
     get dirty() {
-      const currentValue = this.value ? this.value[0] : this.value;
-      return props.value !== currentValue;
+      return props.value !== this.value;
     },
     id: props.id,
     name: props.name,
@@ -80,41 +85,44 @@ const CustomFieldInputCurrency = forwardRef(function CustomFieldInputCurrency(pr
         numberValue = parseInt(valueRef.current.value.replace(/[^0-9-]/g, ''), 10);
       }
 
-      return [numberValue, props.currencyCode];
+      return numberValue;
     },
   }));
-
-  const sharedProps = {
-    className: props.className,
-    id: props.id,
-    label: props.label,
-    name: props.name,
-    placeholder: props.placeholder,
-    readOnly: props.readOnly,
-    required: props.required,
-    tooltip: props.tooltip,
-  };
 
   const formattedNumber = formatValue(input, props.currencyCode);
 
   if (isEditing) {
     return (
-      <CustomFieldInputNumber
-        {...sharedProps}
+      <Number
+        className={props.className}
+        id={props.id}
+        name={props.name}
         onBlur={handleOnBlur}
-        onChange={onChange}
+        onChange={props.onChange}
+        onInvalid={onInvalid}
+        placeholder={props.placeholder}
+        readOnly={props.readOnly}
+        required={props.required}
         ref={numberRef}
         step={currencyMetaData[props.currencyCode].step}
+        validationMessage={props.validationMessage}
+        validationMessageTooltip={props.validationMessageTooltip}
         value={input}
       />
     );
   }
 
   return (
-    <CustomFieldInputText
-      {...sharedProps}
-      errorText={props.errorText}
+    <Input
+      classNames={props.className ? {
+        input: props.className,
+      } : undefined}
+      id={props.id}
+      name={props.name}
       onFocus={handleOnFocus}
+      placeholder={props.placeholder}
+      readOnly={props.readOnly}
+      required={props.required}
       ref={componentRef}
       type="text"
       value={formattedNumber}
@@ -122,36 +130,39 @@ const CustomFieldInputCurrency = forwardRef(function CustomFieldInputCurrency(pr
   );
 });
 
-CustomFieldInputCurrency.propTypes = {
+Money.propTypes = {
   className: PropTypes.string,
   currencyCode: currencyCodeType,
-  errorText: PropTypes.string,
   id: PropTypes.string.isRequired,
-  label: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
+  name: PropTypes.string,
   /**
    * The handler is invoked for every native onchange event.
    * The handler will be invoked with the forwarded ref.
    */
   onChange: PropTypes.func,
+  onInvalid: PropTypes.func,
   placeholder: PropTypes.string,
   readOnly: PropTypes.bool,
   required: PropTypes.bool,
-  tooltip: PropTypes.string,
+  validationMessage: PropTypes.string,
+  validationMessageTooltip: PropTypes.bool,
   value: PropTypes.number,
 };
 
-CustomFieldInputCurrency.defaultProps = {
+Money.defaultProps = {
   className: undefined,
   currencyCode: 'USD',
   errorText: undefined,
   name: undefined,
   onChange: () => {},
+  onInvalid: () => {},
   placeholder: undefined,
   readOnly: false,
   required: false,
   tooltip: undefined,
+  validationMessage: '',
+  validationMessageTooltip: false,
   value: undefined,
 };
 
-export default CustomFieldInputCurrency;
+export default Money;